add last changes with fast commit

+++ src/database/migrations/create_permission_role_table.php
@@ -13,8 +13,8 @@ class CreatePermissionRoleTable extends BaseMigration
         $table->engine = 'InnoDB';
 
         $table->increments('id');
-        $table->integer('permission_id')->unsigned();
-        $table->integer('role_id')->unsigned();
+        $table->bigInteger('permission_id')->unsigned();
+        $table->bigInteger('role_id')->unsigned();
         $table->timestamps();
         $table->foreign('permission_id')
           ->references('id')
