+++ src/database/migrations/create_permissions_table.php
@@ -18,8 +18,8 @@ class CreatePermissionsTabler extends BaseMigration
         $table->string('description');
         $table->boolean('active')->default(true);
         $table->timestamps();
-        $table->integer('created_by')->nullable();
-        $table->integer('updated_by')->nullable();
+        $table->bigInteger('created_by')->nullable();
+        $table->bigInteger('updated_by')->nullable();
         $table->engine = 'InnoDB';
         $table->unique('slug');
         $table->index("name");
