+++ src/database/migrations/leftdatetag_create_role_role_table_rightdatetag.php
@@ -12,8 +12,8 @@ class CreateRoleRoleTablerightdatetag extends BaseMigration
     Schema::create($this->mainTable, function (Blueprint $table) {
     disableForeignKeyConstraints();
       $table->increments('id');
-      $table->integer('domineering_role_id')->unsigned();
-      $table->integer('domined_role_id')->unsigned();
+      $table->bigInteger('domineering_role_id')->unsigned();
+      $table->bigInteger('domined_role_id')->unsigned();
       $table->engine = 'InnoDB';
       $table->foreign('domineering_role_id')
       ->references('id')
